# api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/productivity/task-lists": {
      "get": {
        "operationId": "ProductivityController_taskLists",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      },
      "post": {
        "operationId": "ProductivityController_createTaskList",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskListDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/task-lists/{id}": {
      "patch": {
        "operationId": "ProductivityController_updateTaskList",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskListDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      },
      "delete": {
        "operationId": "ProductivityController_deleteTaskList",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/task-tags": {
      "get": {
        "operationId": "ProductivityController_tags",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      },
      "post": {
        "operationId": "ProductivityController_createTag",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/task-sections": {
      "get": {
        "operationId": "ProductivityController_sections",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      },
      "post": {
        "operationId": "ProductivityController_createSection",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskSectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/task-sections/{id}": {
      "patch": {
        "operationId": "ProductivityController_updateSection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskSectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      },
      "delete": {
        "operationId": "ProductivityController_deleteSection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/tasks": {
      "get": {
        "operationId": "listTasks",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "Productivity"
        ]
      },
      "post": {
        "operationId": "ProductivityController_createTask",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
...
    "/productivity/tasks/reorder": {
      "post": {
        "operationId": "ProductivityController_reorderTasks",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderTasksDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/tasks/{id}": {
      "patch": {
        "operationId": "ProductivityController_updateTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      },
      "delete": {
        "operationId": "ProductivityController_deleteTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/tasks/{id}/complete": {
      "post": {
        "operationId": "ProductivityController_completeTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/tasks/{id}/reopen": {
      "post": {
        "operationId": "ProductivityController_reopenTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/tasks/{id}/cancel": {
      "post": {
        "operationId": "ProductivityController_cancelTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/tasks/{id}/archive": {
      "post": {
        "operationId": "ProductivityController_archiveTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/tasks/{id}/reminders": {
      "post": {
        "operationId": "ProductivityController_createReminder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReminderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/task-reminders/{id}/snooze": {
      "post": {
        "operationId": "ProductivityController_snoozeReminder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnoozeReminderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/task-reminders/{id}/dismiss": {
      "post": {
        "operationId": "ProductivityController_dismissReminder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
...
    "/productivity/notifications/{id}/read": {
      "patch": {
        "operationId": "ProductivityController_markNotificationRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/focus-presets": {
      "get": {
        "operationId": "ProductivityController_presets",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      },
      "post": {
        "operationId": "ProductivityController_createPreset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFocusPresetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
...
    "/productivity/focus-sounds/{id}": {
      "delete": {
        "operationId": "ProductivityController_deleteFocusSound",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      },
      "patch": {
        "operationId": "ProductivityController_updateFocusSound",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/focus-sounds/{soundKey}/preferences": {
      "patch": {
        "operationId": "ProductivityController_updateFocusSoundPreference",
        "parameters": [
          {
            "name": "soundKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/focus-sessions": {
      "post": {
        "operationId": "ProductivityController_startFocus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartFocusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/focus-sessions/{id}/{action}": {
      "post": {
        "operationId": "ProductivityController_focusAction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FocusActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/focus-sessions/{id}/adjust": {
      "patch": {
        "operationId": "ProductivityController_adjustFocus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustFocusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/habits": {
      "get": {
        "operationId": "ProductivityController_habits",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      },
      "post": {
        "operationId": "ProductivityController_createHabit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHabitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/habits/{id}": {
      "patch": {
        "operationId": "ProductivityController_updateHabit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHabitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
...
    "/productivity/habit-occurrences/{id}/check-in": {
      "post": {
        "operationId": "ProductivityController_checkIn",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HabitCheckInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
    "/productivity/habit-occurrences/{id}/skip": {
      "post": {
        "operationId": "ProductivityController_skip",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HabitOccurrenceActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
...
    "/productivity/habits/{id}/stats": {
      "get": {
        "operationId": "ProductivityController_habitStats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Productivity"
        ]
      }
    },
...
  },
  "info": {
    "title": "iTu API",
    "description": "Authoritative REST contract for iTu productivity & learning system",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTaskListDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateTaskListDto": {
        "type": "object",
        "properties": {}
      },
      "CreateTagDto": {
        "type": "object",
        "properties": {}
      },
      "CreateTaskSectionDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateTaskSectionDto": {
        "type": "object",
        "properties": {}
      },
      "CreateTaskDto": {
        "type": "object",
        "properties": {}
      },
      "ReorderTasksDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {}
      },
      "CreateReminderDto": {
        "type": "object",
        "properties": {}
      },
      "SnoozeReminderDto": {
        "type": "object",
        "properties": {}
      },
      "CreateFocusPresetDto": {
        "type": "object",
        "properties": {}
      },
      "StartFocusDto": {
        "type": "object",
        "properties": {}
      },
      "FocusActionDto": {
        "type": "object",
        "properties": {}
      },
      "AdjustFocusDto": {
        "type": "object",
        "properties": {}
      },
      "CreateHabitDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateHabitDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}